jsoninfo: helpers: use a double as parameter for abuf_json_float
Higher precision

+++ lib/jsoninfo/src/olsrd_jsoninfo_helpers.c
@@ -145,10 +145,10 @@ void abuf_json_int(struct autobuf *abuf, const char* key, long value) {
   entrynumber[currentjsondepth]++;
 }
 
-void abuf_json_float(struct autobuf *abuf, const char* key, float value) {
+void abuf_json_float(struct autobuf *abuf, const char* key, double value) {
   abuf_json_insert_comma(abuf);
   abuf_json_new_indent(abuf);
-  abuf_appendf(abuf, "\"%s\": %f", key, (double) value);
+  abuf_appendf(abuf, "\"%s\": %f", key, value);
   entrynumber[currentjsondepth]++;
 }
